requirements.txt
python-dotenv
google-generativeai
langchain-community
PyPDF2
faiss-cpu
streamlit
-e .